{"schema":"decodex/commit/1","summary":"Split worker trace job persistence helpers after strict validation.","authority":"manual"}
